feat: 기본 클래스 생성 및 디렉토리 명 수정
