+++ src/main/java/com/gachtaxi/domain/chat/service/ChatRoomService.java
@@ -1,0 +1,9 @@
+package com.gachtaxi.domain.chat.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class ChatRoomService {
+}
