+++ src/main/java/com/gachtaxi/domain/chat/stomp/StompErrorHandler.java
@@ -1,0 +1,8 @@
+package com.gachtaxi.domain.chat.stomp;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.socket.messaging.StompSubProtocolErrorHandler;
+
+@Configuration
+public class StompErrorHandler extends StompSubProtocolErrorHandler {
+}
