+++ src/main/java/com/gachtaxi/domain/chat/stomp/strategy/StompSubscribeStrategy.java
@@ -1,23 +1,31 @@
-package com.gachtaxi.domain.chat.interceptor.strategy;
+package com.gachtaxi.domain.chat.stomp.strategy;
 
-import lombok.extern.slf4j.Slf4j;
+import com.gachtaxi.domain.chat.exception.ChatSubscribeException;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.MessageChannel;
 import org.springframework.messaging.simp.stomp.StompCommand;
 import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
 import org.springframework.stereotype.Component;
 
-@Slf4j
 @Component
-public class DefaultCommandStrategy implements StompCommandStrategy{
+public class StompSubscribeStrategy implements StompCommandStrategy{
+
+    private static final String SUB_END_POINT = "/sub/chat/room/";
 
     @Override
     public boolean supports(StompCommand command) {
-        return false;
+        return StompCommand.SUBSCRIBE.equals(command);
     }
 
     @Override
     public Message<?> preSend(Message<?> message, StompHeaderAccessor accessor, MessageChannel channel) {
+        String destination = accessor.getDestination();
+
+        if (!destination.startsWith(SUB_END_POINT)) {
+            throw new ChatSubscribeException();
+        }
+
         return message;
     }
 }
+
